Add handle_look_directory_request function and update imports
only 4 fonctions to do
